e2e: node: split out SRIOV utils
TBD
